Login Page(Auth Component) refactored
